Added more correct examples, removed bad ones
